requirements.txt
cassandra-driver==3.24.0
click==7.1.2
Flask==1.1.2
gcloud==0.17.0
geomet==0.2.1.post1
googleapis-common-protos==1.52.0
httplib2==0.19.0
itsdangerous==1.1.0
Jinja2==2.11.3
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==3.0.0
protobuf==4.0.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyparsing==2.4.7
Pyrebase==3.0.27
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7.1
six==1.15.0
Werkzeug==1.0.1
